forge_symposia/requirements.txt
python-dotenv
requests
pyjwt
